+++ .github/workflows/snapshots.yaml
@@ -29,6 +29,4 @@ jobs:
       - run: make toolchain
       - run: make snapshot
         env:
-          CLOUD_PROVIDER: aws
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
