+++ hack/release.sh
@@ -4,7 +4,7 @@ set -euo pipefail
 SNAPSHOT_TAG=$(git describe --tags --always)
 RELEASE_REPO=${RELEASE_REPO:-public.ecr.aws/karpenter}
 
-SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
 source "${SCRIPT_DIR}/release_common.sh"
 
 chart() {
@@ -23,7 +23,6 @@ website() {
     find website/content/en/${RELEASE_VERSION}/*/*/*.yaml -type f | xargs perl -i -p -e "s/preview/${RELEASE_VERSION}/g;"
 }
 
-requireCloudProvider
 authenticate
 buildImages $RELEASE_VERSION
 cosignImages
