--- pkg/controllers/manager.go
@@ -1,61 +1,0 @@
-/*
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package controllers
-
-import (
-	"context"
-	"fmt"
-
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/client-go/rest"
-	controllerruntime "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/healthz"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-)
-
-type GenericControllerManager struct {
-	manager.Manager
-}
-
-// NewManagerOrDie instantiates a controller manager or panics
-func NewManagerOrDie(ctx context.Context, config *rest.Config, options controllerruntime.Options) Manager {
-	newManager, err := controllerruntime.NewManager(config, options)
-	if err != nil {
-		panic(fmt.Sprintf("Failed to create controller newManager, %s", err))
-	}
-	if err := newManager.GetFieldIndexer().IndexField(ctx, &v1.Pod{}, "spec.nodeName", func(o client.Object) []string {
-		return []string{o.(*v1.Pod).Spec.NodeName}
-	}); err != nil {
-		panic(fmt.Sprintf("Failed to setup pod indexer, %s", err))
-	}
-	return &GenericControllerManager{Manager: newManager}
-}
-
-// RegisterControllers registers a set of controllers to the controller manager
-func (m *GenericControllerManager) RegisterControllers(ctx context.Context, controllers ...Controller) Manager {
-	for _, c := range controllers {
-		if err := c.Register(ctx, m); err != nil {
-			panic(err)
-		}
-	}
-	if err := m.AddHealthzCheck("healthz", healthz.Ping); err != nil {
-		panic(fmt.Sprintf("Failed to add health probe, %s", err))
-	}
-	if err := m.AddReadyzCheck("readyz", healthz.Ping); err != nil {
-		panic(fmt.Sprintf("Failed to add ready probe, %s", err))
-	}
-	return m
-}
