+++ pkg/controllers/metrics/pod/suite_test.go
@@ -17,8 +17,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/aws/karpenter/pkg/cloudprovider/fake"
-	"github.com/aws/karpenter/pkg/cloudprovider/registry"
 	"github.com/aws/karpenter/pkg/controllers/metrics/pod"
 	"github.com/aws/karpenter/pkg/test"
 	. "github.com/aws/karpenter/pkg/test/expectations"
@@ -41,8 +39,6 @@ func TestAPIs(t *testing.T) {
 
 var _ = BeforeSuite(func() {
 	env = test.NewEnvironment(ctx, func(e *test.Environment) {
-		cloudProvider := &fake.CloudProvider{}
-		registry.RegisterOrDie(ctx, cloudProvider)
 		controller = pod.NewController(env.Client)
 	})
 	Expect(env.Start()).To(Succeed(), "Failed to start environment")
